# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BotArmy Backend",
    "summary": "Complete backend API for BotArmy multi-agent system with HITL oversight and workflow orchestration",
    "description": "BotArmy POC Backend Services - Multi-Agent Orchestration Platform",
    "contact": {
      "name": "BotArmy Development Team",
      "url": "https://github.com/your-org/bmad"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/hitl/{request_id}": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Request",
        "description": "Get a specific HITL request.",
        "operationId": "get_hitl_request_api_v1_hitl__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HitlRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/{request_id}/history": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Request History",
        "description": "Get the history of a HITL request.",
        "operationId": "get_hitl_request_history_api_v1_hitl__request_id__history_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Hitl Request History Api V1 Hitl  Request Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/project/{project_id}/requests": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Project Hitl Requests",
        "description": "Get all HITL requests for a project.",
        "operationId": "get_project_hitl_requests_api_v1_hitl_project__project_id__requests_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HitlRequestResponse"
                  },
                  "title": "Response Get Project Hitl Requests Api V1 Hitl Project  Project Id  Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hitl/{request_id}/context": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Request Context",
        "description": "Get full context for a HITL request including artifacts and task details.",
        "operationId": "get_hitl_request_context_api_v1_hitl__request_id__context_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HitlContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hitl/pending": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Pending Hitl Requests",
        "description": "Get pending HITL requests with optional project filtering.",
        "operationId": "get_pending_hitl_requests_api_v1_hitl_pending_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HitlRequestResponse"
                  },
                  "title": "Response Get Pending Hitl Requests Api V1 Hitl Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HitlContextResponse": {
        "properties": {
          "hitl_request": {
            "$ref": "#/components/schemas/HitlRequestResponse"
          },
          "task": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "context_artifacts": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Context Artifacts",
            "default": []
          },
          "workflow_context": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Context"
          },
          "project_info": {
            "type": "object",
            "title": "Project Info"
          }
        },
        "type": "object",
        "required": [
          "hitl_request",
          "project_info"
        ],
        "title": "HitlContextResponse",
        "description": "Response model for HITL request context."
      },
...
      "HitlRequestResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "user_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Response"
          },
          "response_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Comment"
          },
          "amended_content": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amended Content"
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "responded_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responded At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "project_id",
          "task_id",
          "question",
          "options",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "HitlRequestResponse",
        "description": "Response model for HITL requests."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}